requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
filelock==3.13.1
fonttools==4.59.0
fsspec==2024.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.8
POT==0.9.5
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycocotools==2.0.10
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.4.0
scipy==1.16.1
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
setuptools==78.1.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
wheel==0.45.1
